requirements.txt
flask
flask-cors
python-dotenv
openai
psycopg2-binary
flask-sqlalchemy
requests
python-dateutil